requirements.txt
fastapi
alembic
asyncpg
sqlalchemy
aiosqlite

pydantic

pydantic-settings
uvicorn

pytest

Pyjwt

cryptography